Update 1.7
Matching Cards from the Mine

code that checks to make sure the clicked card is either one higher or lower than the target card.

converted from the integer IDs in SlotDef.hiddenBy to the actual CardProspectors that have that ID.

Made the game know when its status only needs to be checked
once after each time the player has clicked a card, so the check will be called from the end of Prospector.CardClicked().